requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
pymilvus[milvus-lite]>=2.3.0
milvus-lite>=2.4.0
transformers>=4.35.0
ftfy>=6.1.0
regex>=2023.0.0
torch>=2.0.0
torchvision>=0.15.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
